==> vwl5/VL_Empirical/abgabe/code/2-Panel-Models.qmd <==
# Fixed Effects Models

## Regression

Regression Formulae:

$$
y_{ct} = \beta D_t + \sum_{j=1}^p \gamma_j y_{ct-j} + \alpha_c + \delta_t + \epsilon_{ct}
$$

Formula:

-   $y_{ct}$ = log GDP per capita in country *c* at time *t* = column `y`
-   $D_{ct}$ = Dichotomous measure of democracy = column: `dem`
-   *p* lags of log GDP for control
-   $\alpha$ = country fixed effects
-   $\delta$ = time fixed effects
-   $e$ = error

Estimator: Within Estimator

## Replication

### Imports
```{r}
#| output: false
library(plm)
library(tidyverse)
```

### Load the Data

Dataset avaibable online at: [https://economics.mit.edu/people/faculty/daron-acemoglu/data-archive](https://economics.mit.edu/people/faculty/daron-acemoglu/data-archive)

- Download as .rar Archive File (not openable on MacOS)
- Unpack with [The Unarchiver](https://theunarchiver.com/) (free)
- Open the .dta file with R

```{r}
DDCGdata <- haven::read_dta("../data/DDCGdata.dta")

df <- DDCGdata %>%
  dplyr::select(1:30)
  
rm(DDCGdata)
head(df)
```

Creating Lags of Gdp

```{r}
df <- df %>%
  dplyr::group_by(country_name) %>%
  dplyr::mutate(
    lag1 = lag(y, 1),
    lag2 = lag(y, 2),
    lag3 = lag(y, 3),
    lag4 = lag(y, 4),
    lag5 = lag(y, 5),
    lag6 = lag(y, 6),
    lag7 = lag(y, 7),
    lag8 = lag(y, 8)
  )
```

Making it in a panel dataset

```{r}
pdf = pdata.frame(df, index=c("country_name", "year"))
```


Estimating the Dynamic Linear Model

### Model 1 = 1 Lag

```{r}
model1 <- plm(
  y ~ dem + lag1, #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
summary(model1)
```

### Model 2 = 2 Lags

```{r}
model2 <- plm(
  y ~ dem + lag1 + lag2 , #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
summary(model2)
```

### Model 3 = 4 Lags

```{r}
model3 <- plm(
  y ~ dem + lag1 + lag2 + lag3 + lag4, #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
summary(model3)
```

### Model 4 = 8 Lags

```{r}
model4 <- plm(
  y ~ dem + lag1 + lag2 + lag3 +lag4 +lag5 +lag6+ lag7 +lag8, #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
summary(model4)
```

### Save the Models to a file

```{r}
save(model1, model2, model3, model4, file = "../data/models.RData")
```
