==> vwl5/VL_Empirical/abgabe/code/7-Channels.Models.qmd <==
# Channels of Democracy

Potential Channels mentioned and their respective variables:

-   Log Investment = `loginvpc`
-   Log TFP = `ltfp`
-   Index Economic Reforms: `marketref`
-   Trade Share = `ltrade`
-   Government Expenditure = `lgov`
-   Primary Enrollment = `lprienr`
-   Secondary enrollement = `lsecenr`
-   Child Mortality = `lmort`
-   Dummy Unrest = `unrestn`

## Formula

$$
var_{ct} = \beta D_{ct} + \sum_{j=1}^4 \gamma_j y_{ct-j}+  \sum_{j=1}^4 \delta_j var_{ct-j} +\alpha_c + \theta_t + \epsilon_{ct}
$$
p = 4 lags = preferred specification

Explanation:

-   $var_{ct}$ = explaining variable
-   $y_ct$ = log GDP per capita in country *c* at time *t*
-   $D_{ct}$ = Dichotomous measure of democracy
-   *4* lags of log GDP for control
-   *4* lags of the explaining variable for control
-   $\alpha$ = country fixed effects
-   $\theta$ = time fixed effects
-   $e$ = error


## Imports 

```{r}
#| output: false
library(tidyverse)
library(plm)
```

```{r}
DDCGdata <- haven::read_dta("../data/DDCGdata.dta")

pdf <- pdata.frame(DDCGdata, index=c("country_name", "year"))
rm(DDCGdata)

head(pdf)
```

## Models

### Investment Share

```{r}
model1 <- plm(
  loginvpc ~ dem + plm::lag(y,1:4)+ plm::lag(loginvpc, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```

### TFP

```{r}
model2 <- plm(
  ltfp ~ dem + plm::lag(y,1:4)+ plm::lag(ltfp, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```

### Reforms

```{r}
model3 <- plm(
  marketref ~ dem + plm::lag(y,1:4)+ plm::lag(marketref, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```

### Trade

```{r}
model4 <- plm(
  ltrade2 ~ dem + plm::lag(y,1:4)+ plm::lag(ltrade2, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```
### Government Expenditure

```{r}
model5 <- plm(
  lgov ~ dem + plm::lag(y,1:4)+ plm::lag(lgov, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```


### Primary Enrollment

```{r}
model6 <- plm(
  lprienr ~ dem + plm::lag(y,1:4)+ plm::lag(lprienr, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```

### Secondary Enrollment

```{r}
model7 <- plm(
  lsecenr ~ dem + plm::lag(y,1:4)+ plm::lag(lsecenr, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```

### Child Mortality

```{r}
model8 <- plm(
  lmort ~ dem + plm::lag(y,1:4)+ plm::lag(lmort, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```

### Unrest

```{r}
model9 <- plm(
  unrestn ~ dem + plm::lag(y,1:4)+ plm::lag(unrestn, 1:4), #the equation with country fixed effects
  model="within", #The fixed effects within estimator
  data = pdf, 
  effect = "twoways"
  )
```
## Additional Infos about the Models

```{r}
models <- list(model1, model2, model3, model4, model5, model6, model7, model8, model9)
```


### Long Run

Calculate the long run effects for the different models with this Formula:

$$
\frac{ \hat\beta }{1- \sum_{j=1}^p \hat\gamma_j}
$$


```{r}
lre <- list()

for (i in 1:9) {
  beta_hat <- coef(models[[i]])["dem"]
  gamma_hat <- coef(models[[i]])[6:9]
  
  lre[[i]] <- beta_hat / (1 - sum(gamma_hat))
}

lre <- unlist(lre)
lre <- round(lre, 2)
```
### Persistence

Calculate the persistence for the different models with this Formula:

$$
D_{pers} = \sum_{j=1}^p \hat\gamma_j
$$

```{r}
persistence <- list()

for (i in 1:9) {
  gamma_hat <- coef(models[[i]])[6:9]
  
  persistence[[i]] <- sum(gamma_hat)
}

pers <- unlist(persistence)
pers <- round(pers, 2)
```


## Creating the Table


```{r}
#| output: asis
models <- list(model1, model2, model3, model4, model5, model6, model7, model8, model9)

stargazer::stargazer(
          models,
          type = "html", 
          covariate.labels = c("Democracy"),
          title = "Effects of Democracy on Potential Channels", 
          dep.var.labels = c(
              "Log Investment", "Log TFP", "Index Economic Reforms", 
              "Log Trade Share", "Log of Gov. Exp.","Log Primary Enrollment", 
              "Log Secondary Enrollment", "Log Child Mortality", "Dummy Unrest"
              ), 
          omit.stat = c("f", "ser", "rsq", "adj.rsq"),
          keep = "dem",
          keep.stat = "N",
          add.lines = list(
            c("Long Run Effect", lre),
            c("Persistence", pers)
          )
)
```          
Saving the Table as Tex

```{r}
#| results: asis 
#| message: false
#| warning: false
stargazer::stargazer(
          models,
          type = "latex",
          out = "../data/table_channels.tex",
          covariate.labels = c("Democracy"),
          column.sep.width = "0.01pt",
          title = "Effects of Democracy on Potential Channels", 
          dep.var.labels = c(
              "Log Investment", "Log TFP", "Index Economic Reforms", 
              "Log Trade Share", "Log of Gov. Exp.","Log Primary Enrollment", 
              "Log Secondary Enrollment", "Log Child Mortality", "Dummy Unrest"
              ), 
          omit.stat = c("f", "ser", "rsq", "adj.rsq"),
          keep = "dem",
          keep.stat = "N",
          add.lines = list(
            c("Long Run Effect", lre),
            c("Persistence", pers)
          )
)
```
